feat: Add button background drawable for Android
This commit adds a new XML drawable file, , which defines the background shape and color for buttons in the Android app. The shape is a rounded rectangle with a radius of 50dp, and the color is set to . This drawable can be used to style buttons throughout the app.

chore: Update Android dependencies

This commit updates the Android dependencies in the  file of the  module. It adds the  library with version  as a new implementation dependency. This library provides real-time blur effects for views in the app.

feat: Add button background drawable for Android

This commit adds a new XML drawable file, , which defines the background shape and color for buttons in the Android app. The shape is a rounded rectangle with a radius of 50dp, and the color is set to . This drawable can be used to style buttons throughout the app.

chore: Update Android launcher icons

This commit updates the launcher icons for the Android app. It adds a new XML file, , in the  directory, which defines the adaptive icon for the app. The adaptive icon consists of a background drawable and a foreground drawable. The background drawable is set to , and the foreground drawable is set to .

chore: Update Android launcher icons

This commit updates the launcher icons for the Android app. It adds a new XML file, , in the  directory, which defines the adaptive icon for the app. The adaptive icon consists of a background drawable and a foreground drawable. The background drawable is set to , and the foreground drawable is set to .