feat: Add AppCategoryE enum and UserSettingEntity interface
This commit adds the `AppCategoryE` enum to the `src/core/domain/enums/AppCategoryE.ts` file, which defines the categories of apps. It also adds the `UserSettingEntity` interface to the `src/lesson/domain/entities/UserSettingEntity.ts` file, which represents the user settings entity with properties like `_id`, `childrenId`, `deviceToken`, and `appBlocked`.