chore: Update app name in strings.xml and increase gradle JVM memory limit
This commit updates the app name in the strings.xml file from TBD-Edu to ABeeCi. It also increases the JVM memory limit in the gradle.properties file to -Xmx4096M. These changes improve the branding of the app and enhance its performance by allocating more memory to the Gradle daemon process.