feat: Add blockAppsSystem function to ParentScreen
The blockAppsSystem function is added to the ParentScreen component. It handles blocking apps based on the selectedChild and lesson.blockedListAppsSystem. If the platform is Android, it adds the blocked apps to the locked apps list using the addToLockedApps function. Otherwise, it calls the blockApps function with the selectedChild's ID. This change improves the functionality of the ParentScreen component.